[mlir] Move InlinerInterface from Transforms to Interfaces.
