+++ mlir/lib/Dialect/ControlFlow/IR/CMakeLists.txt
@@ -10,6 +10,7 @@ add_mlir_dialect_library(MLIRControlFlowDialect
   LINK_LIBS PUBLIC
   MLIRArithDialect
   MLIRControlFlowInterfaces
+  MLIRInlinerInterface
   MLIRIR
   MLIRSideEffectInterfaces
   )
