+++ mlir/lib/Dialect/ControlFlow/IR/ControlFlowOps.cpp
@@ -21,8 +21,8 @@
 #include "mlir/IR/PatternMatch.h"
 #include "mlir/IR/TypeUtilities.h"
 #include "mlir/IR/Value.h"
+#include "mlir/Interfaces/InlinerInterface.h"
 #include "mlir/Support/MathExtras.h"
-#include "mlir/Transforms/InliningUtils.h"
 #include "llvm/ADT/APFloat.h"
 #include "llvm/ADT/STLExtras.h"
 #include "llvm/Support/FormatVariadic.h"
