+++ mlir/lib/Dialect/Func/IR/FuncOps.cpp
@@ -18,8 +18,8 @@
 #include "mlir/IR/TypeUtilities.h"
 #include "mlir/IR/Value.h"
 #include "mlir/Interfaces/FunctionImplementation.h"
+#include "mlir/Interfaces/InlinerInterface.h"
 #include "mlir/Support/MathExtras.h"
-#include "mlir/Transforms/InliningUtils.h"
 #include "llvm/ADT/APFloat.h"
 #include "llvm/ADT/MapVector.h"
 #include "llvm/ADT/STLExtras.h"
