+++ mlir/lib/Dialect/Linalg/IR/CMakeLists.txt
@@ -23,6 +23,7 @@ add_mlir_dialect_library(MLIRLinalgDialect
   MLIRDialectUtils
   MLIRFunctionInterfaces
   MLIRInferTypeOpInterface
+  MLIRInlinerInterface
   MLIRIR
   MLIRParser
   MLIRShardingInterface
