+++ mlir/lib/Dialect/MemRef/IR/MemRefDialect.cpp
@@ -9,8 +9,8 @@
 #include "mlir/Conversion/ConvertToLLVM/ToLLVMInterface.h"
 #include "mlir/Dialect/Arith/IR/Arith.h"
 #include "mlir/Dialect/MemRef/IR/MemRef.h"
+#include "mlir/Interfaces/InlinerInterface.h"
 #include "mlir/Interfaces/SideEffectInterfaces.h"
-#include "mlir/Transforms/InliningUtils.h"
 #include <optional>
 
 using namespace mlir;
