+++ mlir/lib/Dialect/Shape/IR/CMakeLists.txt
@@ -19,6 +19,7 @@ add_mlir_dialect_library(MLIRShapeDialect
   MLIRDialect
   MLIRFuncDialect
   MLIRFunctionInterfaces
+  MLIRInlinerInterface
   MLIRInferTypeOpInterface
   MLIRIR
   MLIRSideEffectInterfaces
