+++ mlir/lib/Dialect/UB/IR/CMakeLists.txt
@@ -9,5 +9,6 @@ add_mlir_dialect_library(MLIRUBDialect
   MLIRUBOpsInterfacesIncGen
 
   LINK_LIBS PUBLIC
+  MLIRInlinerInterface
   MLIRIR
   )
