+++ mlir/lib/Interfaces/CMakeLists.txt
@@ -10,6 +10,7 @@ set(LLVM_OPTIONAL_SOURCES
   FunctionInterfaces.cpp
   InferIntRangeInterface.cpp
   InferTypeOpInterface.cpp
+  InlinerInterface.cpp
   LoopLikeInterface.cpp
   MemorySlotInterfaces.cpp
   ParallelCombiningOpInterface.cpp
@@ -63,6 +64,16 @@ add_mlir_library(MLIRFunctionInterfaces
 
 add_mlir_interface_library(InferIntRangeInterface)
 add_mlir_interface_library(InferTypeOpInterface)
+
+add_mlir_library(MLIRInlinerInterface
+  InlinerInterface.cpp
+
+  ADDITIONAL_HEADER_DIRS
+  ${MLIR_MAIN_INCLUDE_DIR}/mlir/Interfaces
+
+  LINK_LIBS PUBLIC
+  MLIRIR
+)
 
 add_mlir_library(MLIRLoopLikeInterface
   LoopLikeInterface.cpp
